wip, reorder and split utils
