+++ src/bootnode.ts
@@ -5,7 +5,8 @@ import PeerId from "peer-id";
 export async function generateBootnodeString(
   key: string,
   ip: string,
-  port: number
+  port: number,
+  useWs: boolean = true
 ): Promise<string> {
   let pair = await libp2pKeys.generateKeyPairFromSeed(
     "Ed25519",
@@ -13,6 +14,6 @@ export async function generateBootnodeString(
     1024
   );
   let peerId: PeerId = await PeerId.createFromPrivKey(pair.bytes);
-  const multiaddress = `/ip4/${ip}/tcp/${port}/ws/p2p/${peerId.toB58String()}`;
+  const multiaddress = `/ip4/${ip}/tcp/${port}/${useWs ? "ws/" : "/"}p2p/${peerId.toB58String()}`;
   return multiaddress;
 }
