+++ src/types.d.ts
@@ -1,4 +1,104 @@
-// Computed Network (internal use)
+// Launch Config, there are used user-input
+// mapped from the json/toml to compute the
+// network config to spawn.
+export interface LaunchConfig {
+  settings: Settings;
+  relaychain: RelayChainConfig;
+  parachains: ParachainConfig[];
+  types: any;
+  hrmpChannels?: HrmpChannelsConfig[];
+  configBasePath: string;
+}
+
+export interface Settings {
+  global_volumes?: GlobalVolume[];
+  bootnode?: boolean;
+  bootnode_domain?: string;
+  timeout: number;
+  grafana?: boolean;
+  telemetry?: boolean;
+  prometheus?: boolean;
+  provider: string;
+}
+
+export interface RelayChainConfig {
+  default_command?: string;
+  default_image?: string;
+  default_resources?: Resources;
+  chain: string;
+  chain_spec_path?: string;
+  chain_spec_command?: string;
+  default_args?: string[];
+  default_overrides?: Override[];
+  nodes?: NodeConfig[];
+  node_groups?: NodeGroupConfig[];
+  total_node_in_groups?: number;
+  genesis?: JSON | ObjectJSON;
+}
+
+export interface NodeConfig {
+  name: string;
+  image?: string;
+  command?: string;
+  commandWithArgs?: string;
+  args?: string[];
+  validator: boolean;
+  env?: envVars[];
+  bootnodes?: string[];
+  overrides?: Override[];
+  add_to_bootnodes?: boolean;
+  resources?: Resources;
+}
+
+export interface NodeGroupConfig {
+  name: string;
+  image?: string;
+  command?: string;
+  args?: string[];
+  env?: envVars[];
+  overrides?: Override[];
+  count: string | number;
+  resources?: Resources;
+}
+
+export interface ParachainConfig {
+  id: number;
+  addToGenesis?: boolean;
+  balance?: number;
+  genesis_wasm_path?: string;
+  genesis_wasm_generator?: string;
+  genesis_state_path?: string;
+  genesis_state_generator?: string;
+  cumulus_based?: boolean;
+  bootnodes?: string[];
+  collator?: CollatorConfig;
+  collator_groups?: CollatorGroupConfig[];
+}
+
+export interface CollatorConfig {
+  image?: string;
+  command?: string;
+  commandWithArgs?: string;
+  name?: string;
+  args?: string[];
+  env?: envVars[];
+}
+
+export interface CollatorGroupConfig {
+  collator: CollatorConfig;
+  count: number;
+}
+
+export interface HrmpChannelsConfig {
+  sender: number;
+  recipient: number;
+  maxCapacity: number;
+  maxMessageSize: number;
+}
+
+
+// Computed Network
+// T
 export interface ComputedNetwork {
   settings: Settings;
   relaychain: {
@@ -29,14 +129,11 @@ export interface Node {
   image: string;
   chain: string;
   chainSpec?: string;
-  port?: number;
-  wsPort?: number;
   validator: boolean;
   args: string[];
   env: envVars[];
   bootnodes: string[];
   zombieRole: "temp" | "node" | "bootnode" | "collator" | "cumulus-collator";
-  initContainers?: object[];
   telemetry?: boolean;
   telemetryUrl: string;
   prometheus?: boolean;
@@ -71,19 +168,19 @@ export interface Parachain {
   collators: Node[];
 }
 
-export interface CollatorNodeConfig {
-  image?: string;
-  command?: string;
-  commandWithArgs?: string;
-  name?: string;
-  args?: string[];
-  initContainers?: object[];
-}
+// export interface CollatorNodeConfig {
+//   image?: string;
+//   command?: string;
+//   commandWithArgs?: string;
+//   name?: string;
+//   args?: string[];
+// }
 
 export interface envVars {
   name: string;
   value: string;
 }
+
 
 export interface ChainSpec {
   name: string;
@@ -107,109 +204,12 @@ export interface ChainSpec {
   };
 }
 
-// Launch Config ( user provided config )
-export interface LaunchConfig {
-  settings: Settings;
-  relaychain: RelayChainConfig;
-  parachains: ParachainConfig[];
-  types: any;
-  hrmpChannels?: HrmpChannelsConfig[];
-  configBasePath: string;
-}
 
-export interface Settings {
-  init_containers?: InitContainer[];
-  global_volumes?: GlobalVolume[];
-  bootnode?: boolean;
-  bootnode_domain?: string;
-  timeout: number;
-  grafana?: boolean;
-  telemetry?: boolean;
-  prometheus?: boolean;
-  provider: string;
-}
-
+// Utils
 export interface GlobalVolume {
   name: string;
   fs_type: string;
   mount_path: string;
-}
-
-export interface InitContainer {
-  image: string;
-  command: string;
-}
-
-export interface RelayChainConfig {
-  default_command?: string;
-  default_image?: string;
-  default_resources?: Resources;
-  chain: string; // rococo-local | local (TODO: move to enum)
-  chain_spec_path?: string;
-  chain_spec_command?: string;
-  default_args?: string[];
-  default_overrides?: Override[];
-  nodes?: NodeConfig[];
-  node_groups?: NodeGroupConfig[];
-  total_node_in_groups?: number;
-  genesis?: JSON | ObjectJSON;
-}
-
-export interface NodeConfig {
-  name: string;
-  image?: string;
-  command?: string;
-  commandWithArgs?: string;
-  wsPort?: number;
-  port?: number;
-  args?: string[];
-  extra_args?: string[];
-  validator: boolean;
-  env?: envVars[];
-  bootnodes?: string[];
-  initContainers?: object[];
-  overrides?: Override[];
-  add_to_bootnodes?: boolean;
-  resources?: Resources;
-}
-
-export interface NodeGroupConfig {
-  name: string;
-  image?: string;
-  command?: string;
-  args?: string[];
-  env?: envVars[];
-  overrides?: Override[];
-  count: string | number;
-  resources?: Resources;
-}
-
-export interface CollatorConfig {
-  image?: string;
-  command?: string;
-  commandWithArgs?: string;
-  name?: string;
-  args?: string[];
-  env?: envVars[];
-}
-
-export interface CollatorGroupConfig {
-  collator: CollatorConfig;
-  count: number;
-}
-
-export interface ParachainConfig {
-  id: number;
-  addToGenesis?: boolean;
-  balance?: number;
-  genesis_wasm_path?: string;
-  genesis_wasm_generator?: string;
-  genesis_state_path?: string;
-  genesis_state_generator?: string;
-  cumulus_based?: boolean;
-  bootnodes?: string[];
-  collator?: CollatorConfig;
-  collator_groups?: CollatorGroupConfig[];
 }
 
 export interface fileMap {
@@ -223,12 +223,7 @@ export interface Override {
   remote_name: string;
 }
 
-export interface HrmpChannelsConfig {
-  sender: number;
-  recipient: number;
-  maxCapacity: number;
-  maxMessageSize: number;
-}
+
 
 interface ObjectJSON {
   [key: string]: ObjectJSON | number | string;
